+++ app/code/community/Lyranetwork/Payzen/Model/Source/PaymentCards.php
@@ -8,6 +8,8 @@
  * @license   https://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  */
 
+use Lyranetwork\Payzen\Model\Api\Form\Api as PayzenApi;
+
 class Lyranetwork_Payzen_Model_Source_PaymentCards
 {
     public function toOptionArray()
@@ -18,7 +20,7 @@ class Lyranetwork_Payzen_Model_Source_PaymentCards
             'label' => Mage::helper('payzen')->__('ALL')
         );
 
-        foreach (Lyranetwork_Payzen_Model_Api_Api::getSupportedCardTypes() as $code => $name) {
+        foreach (PayzenApi::getSupportedCardTypes() as $code => $name) {
             if ($code === 'ONEY_SANDBOX' || $code === 'ONEY') {
                 continue;
             }