+++ CHANGELOG.md
@@ -1,2 +1,11 @@
+1.1.0, 2024-11-22:
+- Added compatibility with OpenMage and PHP latest versions.
+- [technical] Enabled some features by plugin variant.
+- Improve refund management.
+- Update payment method default logo.
+- Update list of supported payment means.
+- Update list of supported currencies.
+- Set return mode to POST by default.
+
 1.0.0, 2021-07-13:
 - Initial plugin version compatible with OpenMage 19 and 20.